+++ tests/roots/test-base/target/faq/inherited_members.py
@@ -1,15 +1,27 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, field_validator
 
 
-class MyBase(BaseModel):
+class Base(BaseModel):
     """MyBase"""
 
     field_on_base: str
     """Base Field"""
 
+    @field_validator("field_on_base")
+    def validate_field_on_base(cls, v):
+        """Validate field_on_base"""
+        return v
 
-class MySubclass(MyBase):
-    """MySubClass"""
+
+class WithoutInheritedMembers(Base):
+    """Without `:inherited-members: BaseModel`"""
+
+    field_on_subclass: str
+    """Subclass field"""
+
+
+class WithInheritedMembers(Base):
+    """With `:inherited-members: BaseModel`"""
 
     field_on_subclass: str
     """Subclass field"""